bot/main.py
import logging
from dotenv import load_dotenv
import os
from telegram.ext import ApplicationBuilder, MessageHandler, filters
from handlers import handle_message

load_dotenv()

logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    level=logging.INFO,
)

def main():
    token = os.getenv("TELEGRAM_BOT_TOKEN")
    app = ApplicationBuilder().token(token).build()

    app.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_message))

    logging.info("Bot started")
    app.run_polling()

if __name__ == "__main__":
    main()
