bot/requirements.txt
python-telegram-bot==21.6
openai==1.82.0
supabase==2.13.0
python-dotenv==1.1.0
